move the migration folder on the script root folder + upodate the corresponding test
