+++ packages/slice-machine/scripts/migration/run.ts
@@ -1,9 +1,9 @@
 import { FileSystem } from "@slicemachine/core";
-import { Migration } from "./migrate";
+import { Migration, MigrationParams } from "./migrate";
 
 export async function run(
   migrations: readonly Migration[],
-  cwd: string
+  params: MigrationParams
 ): Promise<void> {
   if (migrations.length === 0) {
     console.info(
@@ -16,16 +16,16 @@ export async function run(
 
   // run migration
   return head
-    .main(cwd)
+    .main(params)
     .then(() => {
       console.info(
         `Migration ${head.version} done. Read the full changelog for more info!`
       );
       // update last migration version
-      FileSystem.patchManifest(cwd, { _latest: head.version });
+      FileSystem.patchManifest(params.cwd, { _latest: head.version });
 
       // call next migrations
-      return run(tail, cwd);
+      return run(tail, params);
     })
     .catch((e: unknown) => {
       console.error(e);
